Set datastore and base image as separate parameters
Cloning from vCenter templates doesn't require any datastore as
it can look up the correct template by id. I also separated the
parts related to the machine into a separate file and cleaned up
some redundant variables/types
