+++ proto/model.proto
@@ -171,10 +171,15 @@ message EsxiTemplate {
   int32 min_vcpu = 3 [json_name="min_vcpu"];
   int32 min_memory_gb = 4 [json_name="min_memory_gb"];
   message Image {
-    string location = 1 [json_name="location"];
+    message Template {
+      string name = 1 [json_name="name"];
+      string datastore = 2 [json_name="datastore"];
+    }
+    Template template = 1 [json_name="template"];
     string chksum_type = 2 [json_name="chksum_type"];
     string chksum = 3;
   }
+
   Image esxi_image = 5 [json_name="esxi_image"];
   message Interface {
     string type = 1;
@@ -183,6 +188,7 @@ message EsxiTemplate {
   }
   repeated Interface interfaces = 6;
   repeated string node_groups = 7 [json_name="node_groups"];
+  bool use_vcenter = 8 [json_name="use_vcenter"];
 }
 
 message NullTemplate {
